#1153-1 Moved all loop transformations into separate files/directories.
