+++ doc/user_guide/gocean1p0.rst
@@ -885,7 +885,7 @@ the GOcean 1.0 API. For an overview of transformations in general see
 
 ####
 
-.. autoclass:: psyclone.transformations.GOceanLoopFuseTrans
+.. autoclass:: psyclone.domain.gocean.transformations.GOceanLoopFuseTrans
    :members:
    :noindex:
