+++ examples/lfric/eg6/omp_script.py
@@ -38,20 +38,21 @@
 API to apply loop fusion and then OpenMP parallelisation to an invoke
 with two Kernels. This can be applied via the -s option in the
 generator.py script.'''
-from psyclone.transformations import DynamoOMPParallelLoopTrans, \
-    DynamoLoopFuseTrans
+
+from psyclone.configuration import Config
+from psyclone.domain.lfric.transformations import LFRicLoopFuseTrans
+from psyclone.transformations import DynamoOMPParallelLoopTrans
 
 
 def trans(psy):
     ''' PSyclone transformation script for the dynamo0p3 API to apply
     loop fusion and OpenMP for a particular example.'''
     otrans = DynamoOMPParallelLoopTrans()
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
 
     invoke = psy.invokes.invoke_list[0]
     schedule = invoke.schedule
 
-    from psyclone.configuration import Config
     config = Config.get()
     if config.api_conf("dynamo0.3").compute_annexed_dofs and \
        config.distributed_memory:
