+++ src/psyclone/tests/domain/gocean/transformations/gocean1p0_transformations_test.py
@@ -45,18 +45,19 @@ import inspect
 from importlib import import_module
 import pytest
 from psyclone.configuration import Config
+from psyclone.domain.gocean.transformations import GOceanLoopFuseTrans, \
+    GOMoveIterationBoundariesInsideKernelTrans
 from psyclone.undoredo import Memento
 from psyclone.errors import GenerationError
 from psyclone.gocean1p0 import GOLoop
 from psyclone.psyir.nodes import Loop
-from psyclone.psyir.transformations import TransformationError, LoopTrans
+from psyclone.psyir.transformations import LoopFuseTrans, LoopTrans, \
+    TransformationError
 from psyclone.transformations import ACCKernelsTrans, GOConstLoopBoundsTrans, \
-    LoopFuseTrans, GOLoopSwapTrans, OMPParallelTrans, \
+    GOLoopSwapTrans, OMPParallelTrans, \
     GOceanOMPParallelLoopTrans, GOceanOMPLoopTrans, KernelModuleInlineTrans, \
-    GOceanLoopFuseTrans, ACCParallelTrans, ACCEnterDataTrans, ACCLoopTrans, \
+    ACCParallelTrans, ACCEnterDataTrans, ACCLoopTrans, \
     OCLTrans, OMPLoopTrans
-from psyclone.domain.gocean.transformations import \
-    GOMoveIterationBoundariesInsideKernelTrans
 from psyclone.tests.gocean1p0_build import GOcean1p0Build, GOcean1p0OpenCLBuild
 from psyclone.tests.utilities import count_lines, get_invoke, Compile
 
@@ -1526,16 +1527,16 @@ def test_go_loop_swap_wrong_loop_type():
     Test loop swapping transform when supplied loops are not GOLoops.
     '''
     swap = GOLoopSwapTrans()
-    psy, invoke = get_invoke("1.0.1_single_named_invoke.f90",
-                             "dynamo0.3", idx=0, dist_mem=True)
+    _, invoke = get_invoke("1.0.1_single_named_invoke.f90",
+                           "dynamo0.3", idx=0, dist_mem=True)
     with pytest.raises(TransformationError) as error:
         swap.apply(invoke.schedule.children[4])
 
     assert re.search("Given node .* is not a GOLoop, but an instance of "
                      ".*DynLoop", str(error.value)) is not None
 
-    psy, invoke_loop1 = get_invoke("test27_loop_swap.f90", API, idx=1,
-                                   dist_mem=False)
+    _, invoke_loop1 = get_invoke("test27_loop_swap.f90", API, idx=1,
+                                 dist_mem=False)
     schedule = invoke_loop1.schedule
     loop = schedule[0].loop_body[0]
     assert isinstance(loop, GOLoop)
