+++ src/psyclone/tests/domain/lfric/transformations/dynamo0p3_transformations_test.py
@@ -34,6 +34,7 @@
 # Authors R. W. Ford, A. R. Porter and S. Siso, STFC Daresbury Lab
 # Modified I. Kavcic and A. Coughtrie, Met Office
 #          C.M. Maynard, Met Office / University of Reading
+# MOdified J. Henrichs, Bureau of Meteorology
 
 ''' Tests of transformations with the Dynamo 0.3 API '''
 
@@ -41,32 +42,30 @@ from __future__ import absolute_import, print_function
 import pytest
 import inspect
 from importlib import import_module
+
+from psyclone.configuration import Config
 from psyclone.core.access_type import AccessType
-from psyclone import psyGen
-from psyclone import psyir
+from psyclone.domain.lfric.transformations import LFRicLoopFuseTrans
+from psyclone.dynamo0p3 import DynLoop
 from psyclone.errors import GenerationError, InternalError
+from psyclone.psyGen import OMPDoDirective, InvokeSchedule, Directive, \
+    GlobalSum, BuiltIn
+from psyclone.psyir.nodes import colored, Loop, Schedule, Literal
 from psyclone.psyir.symbols import LocalInterface, ScalarType, ArrayType, \
     REAL_TYPE, INTEGER_TYPE
-from psyclone.psyir.nodes import Loop
-from psyclone.psyir.transformations import TransformationError, LoopTrans
+from psyclone.psyir.transformations import LoopFuseTrans, LoopTrans, \
+    TransformationError
 from psyclone.tests.lfric_build import LFRicBuild
 from psyclone.tests.utilities import get_invoke
-from psyclone.transformations import OMPParallelTrans, LoopFuseTrans, \
+from psyclone.transformations import OMPParallelTrans, \
     Dynamo0p3ColourTrans, \
     Dynamo0p3OMPLoopTrans, \
     DynamoOMPParallelLoopTrans, \
-    DynamoLoopFuseTrans, \
     KernelModuleInlineTrans, \
     MoveTrans, \
     Dynamo0p3RedundantComputationTrans, \
     Dynamo0p3AsyncHaloExchangeTrans, \
     Dynamo0p3KernelConstTrans
-from psyclone.configuration import Config
-from psyclone.dynamo0p3 import DynLoop
-from psyclone.psyir.nodes.node import colored
-from psyclone.psyGen import OMPDoDirective, InvokeSchedule, Directive, \
-    GlobalSum, BuiltIn
-from psyclone.psyir.nodes import Loop, Schedule, Literal
 
 
 # The version of the API that the tests in this file
@@ -932,7 +931,7 @@ def test_loop_fuse_invalid_space(monkeypatch):
     monkeypatch.setattr(fspace, "_orig_name", "not_a_space_name")
 
     # Apply transformation and raise the error
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     with pytest.raises(TransformationError) as excinfo:
         _, _ = ftrans.apply(schedule.children[0],
                             schedule.children[1])
@@ -954,7 +953,7 @@ def test_loop_fuse_different_spaces(monkeypatch, dist_mem):
                                TEST_API, name="invoke_0", dist_mem=dist_mem)
         schedule = invoke.schedule
 
-        ftrans = DynamoLoopFuseTrans()
+        ftrans = LFRicLoopFuseTrans()
         ftrans.same_space = same_space
         mtrans = MoveTrans()
         if dist_mem:
@@ -986,7 +985,7 @@ def test_loop_fuse_same_space_error():
     to incorrectly set the 'same_space' property
 
     '''
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     # Assert that 'same_space' is set to default False
     assert ftrans.same_space is False
     # Assert that None is also set to default False
@@ -1010,7 +1009,7 @@ def test_loop_fuse(dist_mem):
     else:
         index = 0
 
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
 
     # Fuse the loops
     schedule, _ = ftrans.apply(schedule.children[index],
@@ -1050,7 +1049,7 @@ def test_loop_fuse_set_dirty():
                              name="invoke_0", dist_mem=True)
 
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     # Fuse the loops
     schedule, _ = ftrans.apply(schedule.children[4],
                                schedule.children[5])
@@ -1070,7 +1069,7 @@ def test_loop_fuse_omp(dist_mem):
                              name="invoke_0", dist_mem=dist_mem)
     schedule = invoke.schedule
 
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     otrans = DynamoOMPParallelLoopTrans()
 
     schedule, _ = ftrans.apply(schedule.children[0],
@@ -1130,7 +1129,7 @@ def test_loop_fuse_omp_rwdisc(tmpdir, monkeypatch, annexed, dist_mem):
                              name="invoke_0", dist_mem=dist_mem)
     schedule = invoke.schedule
 
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     otrans = DynamoOMPParallelLoopTrans()
 
     if dist_mem and not annexed:
@@ -1200,7 +1199,7 @@ def test_fuse_colour_loops(tmpdir, monkeypatch, annexed, dist_mem):
     ctrans = Dynamo0p3ColourTrans()
     otrans = Dynamo0p3OMPLoopTrans()
     rtrans = OMPParallelTrans()
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     mtrans = MoveTrans()
 
     if dist_mem:
@@ -1291,7 +1290,7 @@ def test_loop_fuse_cma(tmpdir, dist_mem):
     '''
     psy, invoke = get_invoke("20.6_multi_invoke_with_cma.f90", TEST_API,
                              name="invoke_0", dist_mem=dist_mem)
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
     mtrans = MoveTrans()
     schedule = invoke.schedule
@@ -1547,7 +1546,7 @@ def test_builtin_loop_fuse_pdo(tmpdir, monkeypatch, annexed, dist_mem):
     psy, invoke = get_invoke("15.14.2_multiple_set_kernels.f90", TEST_API,
                              idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
     schedule, _ = ftrans.apply(schedule.children[0], schedule.children[1])
     schedule, _ = ftrans.apply(schedule.children[0], schedule.children[1])
@@ -1751,7 +1750,7 @@ def test_builtin_loop_fuse_do(tmpdir, monkeypatch, annexed, dist_mem):
     psy, invoke = get_invoke("15.14.2_multiple_set_kernels.f90", TEST_API,
                              idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
     schedule, _ = ftrans.apply(schedule[0], schedule[1])
     schedule, _ = ftrans.apply(schedule[0], schedule[1])
@@ -1894,7 +1893,7 @@ def test_multi_reduction_real_pdo(tmpdir, dist_mem):
     otrans = DynamoOMPParallelLoopTrans()
     # Apply OpenMP parallelisation to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child)
     invoke.schedule = schedule
     code = str(psy.gen)
@@ -1969,7 +1968,7 @@ def test_reduction_after_normal_real_do(tmpdir, monkeypatch, annexed,
     rtrans = OMPParallelTrans()
     # Apply an OpenMP do to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child, {"reprod": False})
     # Apply an OpenMP Parallel for all loops
     schedule, _ = rtrans.apply(schedule.children[0:2])
@@ -2048,7 +2047,7 @@ def test_reprod_red_after_normal_real_do(tmpdir, monkeypatch, annexed,
     rtrans = OMPParallelTrans()
     # Apply an OpenMP do to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child, {"reprod": True})
     # Apply an OpenMP Parallel for all loops
     schedule, _ = rtrans.apply(schedule.children[0:2])
@@ -2149,7 +2148,7 @@ def test_two_reductions_real_do(tmpdir, dist_mem):
     rtrans = OMPParallelTrans()
     # Apply an OpenMP do to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child, {"reprod": False})
     # Apply an OpenMP Parallel for all loops
     schedule, _ = rtrans.apply(schedule.children[0:2])
@@ -2223,7 +2222,7 @@ def test_two_reprod_reductions_real_do(tmpdir, dist_mem):
     rtrans = OMPParallelTrans()
     # Apply an OpenMP do to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child, {"reprod": True})
     # Apply an OpenMP Parallel for all loops
     schedule, _ = rtrans.apply(schedule.children[0:2])
@@ -2325,7 +2324,7 @@ def test_multi_reduction_same_name_real_do():
             rtrans = OMPParallelTrans()
             # Apply an OpenMP do to the loop
             for child in schedule.children:
-                if isinstance(child, psyir.nodes.Loop):
+                if isinstance(child, Loop):
                     schedule, _ = otrans.apply(child, {"reprod": reprod})
             if distmem:
                 # We have to move/delete a
@@ -2354,7 +2353,7 @@ def test_multi_reduction_real_fuse():
                                    idx=0, dist_mem=distmem)
             schedule = invoke.schedule
 
-            ftrans = DynamoLoopFuseTrans()
+            ftrans = LFRicLoopFuseTrans()
             ftrans.same_space = True
             if distmem:
                 # We need to remove the global sum. This makes the
@@ -2364,7 +2363,7 @@ def test_multi_reduction_real_fuse():
             with pytest.raises(TransformationError) as excinfo:
                 schedule, _ = ftrans.apply(schedule.children[0],
                                            schedule.children[1])
-            assert ("Error in DynamoLoopFuseTrans transformation: Cannot "
+            assert ("Error in LFRicLoopFuseTrans transformation: Cannot "
                     "fuse loops when each loop already contains a "
                     "reduction" in str(excinfo.value))
 
@@ -2380,7 +2379,7 @@ def test_multi_different_reduction_real_pdo(tmpdir, dist_mem):
     otrans = DynamoOMPParallelLoopTrans()
     # Apply OpenMP parallelisation to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child)
     invoke.schedule = schedule
     code = str(psy.gen)
@@ -2454,7 +2453,7 @@ def test_multi_builtins_red_then_pdo(tmpdir, monkeypatch, annexed, dist_mem):
     otrans = DynamoOMPParallelLoopTrans()
     # Apply OpenMP parallelisation to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child)
     invoke.schedule = schedule
     result = str(psy.gen)
@@ -2525,7 +2524,7 @@ def test_multi_builtins_red_then_do(tmpdir, monkeypatch, annexed, dist_mem):
     rtrans = OMPParallelTrans()
     # Apply an OpenMP do to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child, {"reprod": False})
     if dist_mem:  # annexed can be True or False
         mtrans = MoveTrans()
@@ -2603,7 +2602,7 @@ def test_multi_builtins_red_then_fuse_pdo(tmpdir, monkeypatch, annexed,
     file_name = "15.17.1_one_reduction_one_standard_builtin.f90"
     psy, invoke = get_invoke(file_name, TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
     if dist_mem and annexed:
         mtrans = MoveTrans()
@@ -2681,7 +2680,7 @@ def test_multi_builtins_red_then_fuse_do(tmpdir, monkeypatch, annexed,
     file_name = "15.17.1_one_reduction_one_standard_builtin.f90"
     psy, invoke = get_invoke(file_name, TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
     if dist_mem and annexed:
         mtrans = MoveTrans()
@@ -2762,7 +2761,7 @@ def test_multi_builtins_usual_then_red_pdo(tmpdir, monkeypatch, annexed,
     otrans = DynamoOMPParallelLoopTrans()
     # Apply OpenMP parallelisation to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child)
     invoke.schedule = schedule
     result = str(psy.gen)
@@ -2834,7 +2833,7 @@ def test_builtins_usual_then_red_fuse_pdo(tmpdir, monkeypatch, annexed,
     file_name = "15.17.2_one_standard_builtin_one_reduction.f90"
     psy, invoke = get_invoke(file_name, TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
 
     if dist_mem and annexed:
@@ -2903,7 +2902,7 @@ def test_builtins_usual_then_red_fuse_do(tmpdir, monkeypatch, annexed,
     file_name = "15.17.2_one_standard_builtin_one_reduction.f90"
     psy, invoke = get_invoke(file_name, TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
 
     if dist_mem and annexed:
@@ -2975,7 +2974,7 @@ def test_multi_builtins_fuse_error():
     _, invoke = get_invoke("15.18.1_builtins_reduction_fuse_error.f90",
                            TEST_API, idx=0, dist_mem=False)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
     schedule, _ = ftrans.apply(schedule.children[0], schedule.children[1])
     with pytest.raises(TransformationError) as excinfo:
@@ -2991,7 +2990,7 @@ def test_loop_fuse_error(dist_mem):
     _, invoke = get_invoke("15.14.2_multiple_set_kernels.f90",
                            TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     with pytest.raises(TransformationError) as excinfo:
         schedule, _ = ftrans.apply(schedule.children[0],
                                    schedule.children[1])
@@ -3098,7 +3097,7 @@ def test_no_global_sum_in_parallel_region():
         rtrans = OMPParallelTrans()
         # Apply an OpenMP do to the loop
         for child in schedule.children:
-            if isinstance(child, psyir.nodes.Loop):
+            if isinstance(child, Loop):
                 schedule, _ = otrans.apply(child, {"reprod": True})
         schedule, _ = rtrans.apply(schedule.children)
         invoke.schedule = schedule
@@ -3126,7 +3125,7 @@ def test_reprod_builtins_red_then_usual_do(tmpdir, monkeypatch, annexed,
     rtrans = OMPParallelTrans()
     # Apply an OpenMP do to the loop
     for child in schedule.children:
-        if isinstance(child, psyir.nodes.Loop):
+        if isinstance(child, Loop):
             schedule, _ = otrans.apply(child, {"reprod": True})
     if dist_mem:  # annexed can be True or False
         mtrans = MoveTrans()
@@ -3236,7 +3235,7 @@ def test_repr_bltins_red_then_usual_fuse_do(tmpdir, monkeypatch, annexed,
     file_name = "15.17.1_one_reduction_one_standard_builtin.f90"
     psy, invoke = get_invoke(file_name, TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
 
     if dist_mem:  # annexed can be True or False
@@ -3350,7 +3349,7 @@ def test_repr_bltins_usual_then_red_fuse_do(tmpdir, monkeypatch, annexed,
     file_name = "15.17.2_one_standard_builtin_one_reduction.f90"
     psy, invoke = get_invoke(file_name, TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     ftrans.same_space = True
 
     if dist_mem and annexed:
@@ -3433,8 +3432,6 @@ def test_repr_3_builtins_2_reductions_do(tmpdir, dist_mem):
     ''' Test that we generate correct reproducible OpenMP DO reductions
     when we have three different builtins, first a reduction, then a
     normal builtin then a reduction. '''
-    from psyclone.dynamo0p3 import DynLoop
-    from psyclone.psyGen import OMPDoDirective
     file_name = "15.19.1_three_builtins_two_reductions.f90"
     psy, invoke = get_invoke(file_name, TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
@@ -5030,10 +5027,10 @@ def test_loop_fusion_different_loop_depth(monkeypatch, annexed):
         # we now have an additional halo exchange for the gh_inc
         # access in the first loop
         index += 1
-    f_trans = DynamoLoopFuseTrans()
+    f_trans = LFRicLoopFuseTrans()
     with pytest.raises(TransformationError) as excinfo:
         f_trans.apply(schedule.children[index], schedule.children[index+1])
-    assert ("Error in DynamoLoopFuseTrans transformation: The halo-depth "
+    assert ("Error in LFRicLoopFuseTrans transformation: The halo-depth "
             "indices are not the same. Found '3' and '1'" in
             str(excinfo.value))
     # now redundantly compute to the full halo
@@ -5046,10 +5043,10 @@ def test_loop_fusion_different_loop_depth(monkeypatch, annexed):
         move_trans = MoveTrans()
         move_trans.apply(schedule.children[7], schedule.children[6])
     # try to fuse the loops. This should fail as the depths are different
-    f_trans = DynamoLoopFuseTrans()
+    f_trans = LFRicLoopFuseTrans()
     with pytest.raises(TransformationError) as excinfo:
         f_trans.apply(schedule.children[index], schedule.children[index+1])
-    assert ("Error in DynamoLoopFuseTrans transformation: The halo-depth "
+    assert ("Error in LFRicLoopFuseTrans transformation: The halo-depth "
             "indices are not the same. Found '3' and 'None'" in
             str(excinfo.value))
 
@@ -5065,11 +5062,11 @@ def test_loop_fusion_different_loop_name(monkeypatch):
     schedule = invoke.schedule
     rc_trans = Dynamo0p3RedundantComputationTrans()
     rc_trans.apply(schedule.children[0], {"depth": 3})
-    f_trans = DynamoLoopFuseTrans()
+    f_trans = LFRicLoopFuseTrans()
     with pytest.raises(TransformationError) as excinfo:
         # Indices of loops to fuse in the schedule
         f_trans.apply(schedule.children[2], schedule.children[3])
-    assert ("Error in DynamoLoopFuseTrans transformation: The upper bound "
+    assert ("Error in LFRicLoopFuseTrans transformation: The upper bound "
             "names are not the same. Found 'cell_halo' and 'ncells'"
             in str(excinfo.value))
     # Now test for f1 write to read dependency
@@ -5084,7 +5081,7 @@ def test_loop_fusion_different_loop_name(monkeypatch):
     rc_trans.apply(schedule.children[0], {"depth": 3})
     with pytest.raises(TransformationError) as excinfo:
         f_trans.apply(schedule.children[1], schedule.children[2])
-    assert ("Error in DynamoLoopFuseTrans transformation: The upper bound "
+    assert ("Error in LFRicLoopFuseTrans transformation: The upper bound "
             "names are not the same. Found 'cell_halo' and 'ncells'"
             in str(excinfo.value))
 
@@ -5617,7 +5614,7 @@ def test_loop_fuse_then_rc(tmpdir):
                              TEST_API, name="invoke_0", dist_mem=True)
     schedule = invoke.schedule
 
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
 
     # Fuse the loops
     schedule, _ = ftrans.apply(schedule.children[4],
@@ -6117,7 +6114,7 @@ def test_intergrid_colour(dist_mem):
                              TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
     # First two kernels are prolongation, last two are restriction
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     ctrans = Dynamo0p3ColourTrans()
     # To a prolong kernel
     _, _ = ctrans.apply(loops[1])
@@ -6160,12 +6157,12 @@ def test_intergrid_colour_errors(dist_mem, monkeypatch):
                              TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
     # First two kernels are prolongation, last two are restriction
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     loop = loops[1]
     # To a prolong kernel
     _, _ = ctrans.apply(loop)
     # Update our list of loops
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     # Trigger the error by calling the internal method to get the upper
     # bound before the colourmaps have been set-up
     with pytest.raises(InternalError) as err:
@@ -6198,13 +6195,13 @@ def test_intergrid_omp_parado(dist_mem, tmpdir):
                              TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
     # First two kernels are prolongation, last two are restriction
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     ctrans = Dynamo0p3ColourTrans()
     # To a prolong kernel
     _, _ = ctrans.apply(loops[1])
     # To a restrict kernel
     _, _ = ctrans.apply(loops[3])
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     otrans = DynamoOMPParallelLoopTrans()
     # Apply OMP to loops over coloured cells
     _, _ = otrans.apply(loops[2])
@@ -6233,13 +6230,13 @@ def test_intergrid_omp_para_region1(dist_mem, tmpdir):
     ptrans = OMPParallelTrans()
     otrans = Dynamo0p3OMPLoopTrans()
     # Colour the first loop
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     _, _ = ctrans.apply(loops[0])
     # Parallelise the loop over cells of a given colour
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     _, _ = otrans.apply(loops[1])
     # Put the parallel loop inside a parallel region
-    dirs = schedule.walk(psyGen.Directive)
+    dirs = schedule.walk(Directive)
     _, _ = ptrans.apply(dirs[0])
     gen = str(psy.gen)
     if dist_mem:
@@ -6271,13 +6268,13 @@ def test_intergrid_omp_para_region2(dist_mem, tmpdir):
                              TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
     schedule.view()
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     ctrans = Dynamo0p3ColourTrans()
-    ftrans = DynamoLoopFuseTrans()
+    ftrans = LFRicLoopFuseTrans()
     _, _ = ctrans.apply(loops[0])
     _, _ = ctrans.apply(loops[1])
     schedule.view()
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     _, _ = ftrans.apply(loops[0], loops[2])
     schedule.view()
     assert LFRicBuild(tmpdir).code_compiles(psy)
@@ -6292,7 +6289,7 @@ def test_intergrid_err(dist_mem):
                            TEST_API, idx=0, dist_mem=dist_mem)
     schedule = invoke.schedule
     # First two kernels are prolongation, last two are restriction
-    loops = schedule.walk(psyir.nodes.Loop)
+    loops = schedule.walk(Loop)
     expected_err = (
         "cannot currently be applied to nodes which have inter-grid "
         "kernels as descendents and ")
@@ -6304,7 +6301,7 @@ def test_intergrid_err(dist_mem):
             rc_trans.apply(loops[2], {"depth": 2})
             assert expected_err in str(excinfo.value)
 
-    lftrans = DynamoLoopFuseTrans()
+    lftrans = LFRicLoopFuseTrans()
     with pytest.raises(TransformationError) as excinfo:
         lftrans.apply(loops[0], loops[1])
     assert expected_err in str(excinfo.value)
