+++ src/psyclone/tests/dynamo0p3_test.py
@@ -58,11 +58,11 @@ from psyclone.dynamo0p3 import DynKernMetadata, DynKern, \
     DynLoop, DynGlobalSum, HaloReadAccess, \
     KernCallArgList, DynACCEnterDataDirective, VALID_INTRINSIC_TYPES
 
-from psyclone.transformations import LoopFuseTrans
 from psyclone.gen_kernel_stub import generate
 from psyclone.configuration import Config
 from psyclone.tests.lfric_build import LFRicBuild
 from psyclone.psyir.nodes import Schedule
+from psyclone.psyir.transformations import LoopFuseTrans
 
 # constants
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
