+++ src/psyclone/tests/psyir/transformations/transformations_test.py
@@ -33,6 +33,7 @@
 # ----------------------------------------------------------------------------
 # Authors R. W. Ford, A. R. Porter and S. Siso, STFC Daresbury Lab
 # Modified I. Kavcic, Met Office
+# Modified J. Henrichs, Bureau of Metorology
 
 '''
 API-agnostic tests for various transformation classes.
@@ -40,19 +41,21 @@ API-agnostic tests for various transformation classes.
 
 from __future__ import absolute_import, print_function
 import pytest
-from psyclone.psyir.nodes import Literal, Loop, Node, Reference, Schedule, \
-    Statement, CodeBlock
+
+from psyclone.errors import InternalError
+from psyclone.psyGen import ACCLoopDirective
+from psyclone.psyir.nodes import IfBlock, Literal, Loop, Node, Reference, \
+    Schedule, Statement, CodeBlock, Return
 from psyclone.psyir.symbols import DataSymbol, INTEGER_TYPE, BOOLEAN_TYPE
-from psyclone.psyir.transformations import ProfileTrans, RegionTrans, \
-    TransformationError
-from psyclone.transformations import ACCParallelTrans, OMPParallelLoopTrans, \
-    OMPParallelTrans
+from psyclone.psyir.transformations import LoopFuseTrans, ProfileTrans, \
+    RegionTrans, TransformationError
+from psyclone.tests.utilities import get_invoke
+from psyclone.transformations import ACCEnterDataTrans, ACCLoopTrans, \
+    ACCParallelTrans, OMPLoopTrans, OMPParallelLoopTrans, OMPParallelTrans
 
 
 def test_accloop():
     ''' Generic tests for the ACCLoopTrans transformation class '''
-    from psyclone.transformations import ACCLoopTrans
-    from psyclone.psyGen import ACCLoopDirective
     trans = ACCLoopTrans()
     assert trans.name == "ACCLoopTrans"
     assert str(trans) == "Adds an 'OpenACC loop' directive to a loop"
@@ -71,7 +74,6 @@ def test_accparallel():
 
 def test_accenterdata():
     ''' Generic tests for the ACCEnterDataTrans class '''
-    from psyclone.transformations import ACCEnterDataTrans
     acct = ACCEnterDataTrans()
     assert acct.name == "ACCEnterDataTrans"
     assert str(acct) == "Adds an OpenACC 'enter data' directive"
@@ -81,8 +83,6 @@ def test_accenterdata_internalerr(monkeypatch):
     ''' Check that the ACCEnterDataTrans.apply() method raises an internal
     error if the validate method fails to throw out an invalid type of
     Schedule. '''
-    from psyclone.transformations import ACCEnterDataTrans
-    from psyclone.errors import InternalError
     acct = ACCEnterDataTrans()
     monkeypatch.setattr(acct, "validate", lambda sched, options: None)
     with pytest.raises(InternalError) as err:
@@ -94,7 +94,6 @@ def test_accenterdata_internalerr(monkeypatch):
 def test_omploop_no_collapse():
     ''' Check that the OMPLoopTrans.directive() method rejects the
     collapse argument '''
-    from psyclone.transformations import OMPLoopTrans
     trans = OMPLoopTrans()
     pnode = Node()
     cnode = Node()
@@ -108,7 +107,6 @@ def test_ifblock_children_region():
     ''' Check that we reject attempts to transform the conditional part of
     an If statement or to include both the if- and else-clauses in a region
     (without their parent). '''
-    from psyclone.psyir.nodes import IfBlock
     acct = ACCParallelTrans()
     # Construct a valid IfBlock
     condition = Reference(DataSymbol('condition', BOOLEAN_TYPE))
@@ -129,8 +127,6 @@ def test_ifblock_children_region():
 
 def test_fusetrans_error_incomplete():
     ''' Check that we reject attempts to fuse loops which are incomplete. '''
-    from psyclone.psyir.nodes import Return
-    from psyclone.transformations import LoopFuseTrans
     sch = Schedule()
     loop1 = Loop(variable=DataSymbol("i", INTEGER_TYPE), parent=sch)
     loop2 = Loop(variable=DataSymbol("j", INTEGER_TYPE), parent=sch)
@@ -170,7 +166,6 @@ def test_fusetrans_error_incomplete():
 def test_fusetrans_error_not_same_parent():
     ''' Check that we reject attempts to fuse loops which don't share the
     same parent '''
-    from psyclone.transformations import LoopFuseTrans
 
     sch1 = Schedule()
     sch2 = Schedule()
@@ -268,7 +263,6 @@ def test_profile_trans_name(options):
     set to None.
 
     '''
-    from psyclone.tests.utilities import get_invoke
     _, invoke = get_invoke("1_single_invoke.f90", "dynamo0.3", idx=0)
     schedule = invoke.schedule
     profile_trans = ProfileTrans()
